Create Palindromes
Palindromes
This is an example of a program that checks whether or not a piece of text is a palindrome. A palindrome is a string whose letters read the same backwards and forwards. For example, 'Kayak` and 'A man, a plan, a canal - Panama!' are both palindromes. 

We can check whether a string is a Palindrome by reversing it, and then checking if it is equal to its reversed version.

There are two functions of particular interest in this program:

normalize, which accepts a string parameter and returns a new string consisting of only the letters in the first string

reverse_string, which accepts a string parameter and returns the reverse of that string