Create MPedigree
This is an example program that shows how MPedigree's goldkeys system might work to produce label numbers for boxes of medicine. The program produces N_LABELS labels, each of which is a 5-digit random number followed a 4-digit unique number. 

In particular, the pad function takes in an integer num and a length length as a parameter and produces a string version of that integer that is length characters long, padding with 0s if need be.